move compute functions
