+++ store/buildPlayer.js
@@ -5,9 +5,9 @@ var async = require('async');
 var constants = require('../constants.js');
 var queries = require("../store/queries");
 var utility = require('../util/utility');
+var aggregator = require('../util/aggregator');
 var config = require('../config');
 var playerCache = require('../store/playerCache');
-var aggregator = require('../compute/aggregator');
 var getPlayerMatches = queries.getPlayerMatches;
 var getPlayer = queries.getPlayer;
 var getPlayerRankings = queries.getPlayerRankings;
